Update to mcrouter v2024.06.10
Bump mcrouter to v2024.06.10. I have not been able to find a combination
of FB OSS tags that would work together in the last few weeks because
every time the OSS build breaks in one of the libraries, so give up and
use a newer tag for fbthrift to get it to build. Patch mcrouter too
because the OSS build has grown stale (this will be upstreamed).

Also override `dh_strip` to preserve debug info in mcrouter binaries --
which could allow us to profile them later if we so desire.